Classification/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.2.0
cycler==0.12.1
filelock==3.9.0
fonttools==4.44.0
fsspec==2023.4.0
idna==3.4
imageio==2.32.0
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.0

packaging==23.2
pandas==2.1.3
Pillow==9.3.0
pyparsing==3.1.1
PySnooper==1.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
SimpleITK==2.3.1
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
TorchSnooper==0.8
numpy==1.24.1
opencv-python==4.8.1.78
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
zipp==3.17.0
openpyxl